+++ src/main/java/kdt/shorturl/url/converter/ShortUrlConverter.java
@@ -1,4 +1,4 @@
-package kdt.shorturl.grobal.util;
+package kdt.shorturl.url.converter;
 
 public interface ShortUrlConverter {
     String encoding(int num);
